+++ src/main/java/uk/gov/companieshouse/chs/notification/sender/api/mapper/NotificationMapper.java
@@ -1,16 +1,16 @@
 package uk.gov.companieshouse.chs.notification.sender.api.mapper;
 
-import org.mapstruct.Mapping;
-import org.mapstruct.Named;
-import org.mapstruct.Mapper;
-import org.springframework.stereotype.Component;
-import uk.gov.companieshouse.api.chs_notification_sender.model.GovUkEmailDetailsRequest;
-import uk.gov.companieshouse.api.chs_notification_sender.model.GovUkLetterDetailsRequest;
-import uk.gov.companieshouse.notification.ChsEmailNotification;
-import uk.gov.companieshouse.notification.ChsLetterNotification;
-
 import java.time.Instant;
 import java.time.OffsetDateTime;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Named;
+import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.chs.notification.model.GovUkEmailDetailsRequest;
+import uk.gov.companieshouse.api.chs.notification.model.GovUkLetterDetailsRequest;
+import uk.gov.companieshouse.notification.ChsEmailNotification;
+import uk.gov.companieshouse.notification.ChsLetterNotification;
 
 @Component
 @Mapper(componentModel = "spring")
