+++ src/test/java/uk/gov/companieshouse/chs/notification/sender/api/controller/NotificationSenderControllerTest.java
@@ -18,11 +18,11 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import uk.gov.companieshouse.api.chs_notification_sender.model.GovUkEmailDetailsRequest;
-import uk.gov.companieshouse.api.chs_notification_sender.model.GovUkLetterDetailsRequest;
-import uk.gov.companieshouse.api.chs_notification_sender.model.LetterDetails;
-import uk.gov.companieshouse.api.chs_notification_sender.model.RecipientDetailsLetter;
-import uk.gov.companieshouse.api.chs_notification_sender.model.SenderDetails;
+import uk.gov.companieshouse.api.chs.notification.model.GovUkEmailDetailsRequest;
+import uk.gov.companieshouse.api.chs.notification.model.GovUkLetterDetailsRequest;
+import uk.gov.companieshouse.api.chs.notification.model.LetterDetails;
+import uk.gov.companieshouse.api.chs.notification.model.RecipientDetailsLetter;
+import uk.gov.companieshouse.api.chs.notification.model.SenderDetails;
 import uk.gov.companieshouse.chs.notification.sender.api.TestUtil;
 import uk.gov.companieshouse.chs.notification.sender.api.exception.NotificationException;
 import uk.gov.companieshouse.chs.notification.sender.api.kafka.KafkaProducerService;
